Create LanguageAcquisition
This is a simple program that I worked on for a very minimal amount of time. It's written in Python 2.7.6 64-bit (as all of my programs will be written in) and it's open for interpretation. The reason I put this together today is to demonstrate one method of data integration. It is context sensitive and dispositionally sensitive, but it's not where I want it to be. I put a part of it in triple quotes because that is another aspect of the program that will use statistics to make a prediction of what patterns to use (in this case, it should combine multiple patterns), but I haven't started integrating the two bits of code. Anyways, time is an empty dictionary when you start, and if you want to see how it can really interact, you're probably going to want to keep interacting with it and make sure to restart multiple times so it can get used to the different attitudes that it will have towards certain circumstances. Once time gets large enough, it's going to be a lot more fun so make sure you import cPickle and cPickle.dump(time, open('time.p','wb))